# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movie Picker",
    "description": "Приложение для выбора фильмов с рекомендациями на основе AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Search Movies",
        "description": "Поиск фильмов в OMDB по названию",
        "operationId": "search_movies_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Поисковый запрос",
              "title": "Q"
            },
            "description": "Поисковый запрос"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OMDBSearchResult"
                  },
                  "title": "Response Search Movies Api Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/instagram/import": {
      "post": {
        "tags": [
          "instagram"
        ],
        "summary": "Import From Instagram",
        "description": "Импорт фильмов из Instagram Reel по ссылке.",
        "operationId": "import_from_instagram_api_instagram_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstagramImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  },
                  "type": "array",
                  "title": "Response Import From Instagram Api Instagram Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/instagram/search": {
      "post": {
        "tags": [
          "instagram"
        ],
        "summary": "Search From Instagram",
        "description": "Извлечь названия фильмов из Instagram Reel и найти их в OMDB.",
        "operationId": "search_from_instagram_api_instagram_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstagramImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OMDBSearchResult"
                  },
                  "type": "array",
                  "title": "Response Search From Instagram Api Instagram Search Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InstagramImportRequest": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url"
          },
          "vision": {
            "type": "boolean",
            "title": "Vision",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "url"
        ],
        "title": "InstagramImportRequest",
        "description": "Запрос на импорт фильма(ов) из Instagram Reel по ссылке."
      },
      "Movie": {
        "properties": {
          "imdb_id": {
            "type": "string",
            "title": "Imdb Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres",
            "default": []
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "plot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot"
          },
          "cast": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Cast",
            "default": []
          },
          "director": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Director"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "imdb_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imdb Rating"
          },
          "awards": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Awards"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "is_watched": {
            "type": "boolean",
            "title": "Is Watched",
            "default": false
          },
          "source": {
            "type": "string",
            "title": "Source",
            "default": "personal"
          },
          "added_at": {
            "type": "string",
            "format": "date-time",
            "title": "Added At"
          }
        },
        "type": "object",
        "required": [
          "imdb_id",
          "title",
          "id",
          "added_at"
        ],
        "title": "Movie",
        "description": "Фильм из базы данных"
      },
...
      "OMDBSearchResult": {
        "properties": {
          "imdb_id": {
            "type": "string",
            "title": "Imdb Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "year": {
            "type": "string",
            "title": "Year"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          }
        },
        "type": "object",
        "required": [
          "imdb_id",
          "title",
          "year"
        ],
        "title": "OMDBSearchResult",
        "description": "Результат поиска в OMDB"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}